Updating go modules
